+++ diet/src/main/java/com/example/diet/Service/Impl/RecordServiceImpl.java
@@ -83,14 +83,14 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     @InvokeLog
-    public Long getCal(String userId,String date){
+    public Integer getCal(String userId,String date){
         return recordMapper.getCal(Integer.parseInt(userId),date);
     }
 
 
     @Override
     @InvokeLog
-    public Long getStep(String userId,String date){
+    public Integer getStep(String userId,String date){
         return recordMapper.getStep(Integer.parseInt(userId),date);
     }
